change usr_fldbase, fld_x, and flx_y from unsigned int to size_t

+++ randomart.c
@@ -44,10 +44,10 @@
 #define	FLDSIZE_X	(FLDBASE * 2 + 1)
 
 /* function prototypes */
-char *fingerprint_randomart(char *userstr, size_t userstr_len, unsigned int usr_fldbase);
+char *fingerprint_randomart(char *userstr, size_t userstr_len, size_t usr_fldbase);
 
 char * 
-fingerprint_randomart(char *userstr, size_t userstr_len, unsigned int usr_fldbase) {
+fingerprint_randomart(char *userstr, size_t userstr_len, size_t usr_fldbase) {
 	/*
 	 * Chars to be used after each other every time the worm
 	 * intersects with itself.  Matter of taste.
@@ -60,7 +60,7 @@ fingerprint_randomart(char *userstr, size_t userstr_len, unsigned int usr_fldbas
 	int	x, y;
 //	int	r;
 	size_t	len = strlen(augmentation_string) - 1;
-	unsigned int	fld_x, fld_y;
+	size_t fld_x, fld_y;
 	int	i;
 
 	fld_y = usr_fldbase + 1;
@@ -218,10 +218,10 @@ main(int argc, char **argv)
 	size_t	line_buf_len = 0;
 	ssize_t	line_len;
 	char	*randomart = NULL;
-	unsigned int	usr_fldbase;
+	size_t	usr_fldbase;
 
 	if (argc > 1) {
-		usr_fldbase = (unsigned int)strtoul(argv[1], NULL, 10);
+		usr_fldbase = (size_t)strtoul(argv[1], NULL, 10);
 	} else {
 		usr_fldbase = 8;
 	}
